pyproject.toml
[project]
name = "eight-rack-ai"
version = "0.1.0"
description = "MTG 8 Rack deck pilot and tuning system"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "anthropic>=0.40",
    "instructor>=1.0",
    "openai>=1.0",
    "scrython>=1.9",
    "aiosqlite>=0.20",
    "rich>=13.0",
    "typer>=0.12",
    "pyyaml>=6.0",
    "httpx>=0.27",
]

[project.scripts]
eight-rack = "eight_rack.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/eight_rack"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]
